requirements.txt
flask
flask-cors
numpy
tensorflow-cpu==2.10.0
opencv-python-headless
mediapipe
pillowgunicorn 
gunicorn 
gunicorn 
